metrics is a separate cmake module
