+++ src/CMakeLists.txt
@@ -14,12 +14,13 @@
 
 add_subdirectory(third_party)
 
+add_subdirectory(metrics)
 add_subdirectory(gennylib)
 add_subdirectory(cast_core)
 add_subdirectory(driver)
 
 # This makes the project importable from the build directory
-export(TARGETS gennylib gennylib_cast gennylib_generators gennylib_core loki cast_core
+export(TARGETS metrics gennylib gennylib_cast gennylib_generators gennylib_core loki cast_core
        FILE GennyLibraryConfig.cmake)
 
 # Below generates a cmake module so if you make install from this
