+++ src/gennylib/include/gennylib/context.hpp
@@ -35,8 +35,8 @@
 #include <gennylib/InvalidConfigurationException.hpp>
 #include <gennylib/Orchestrator.hpp>
 #include <gennylib/conventions.hpp>
-#include <gennylib/metrics.hpp>
 #include <gennylib/v1/ConfigNode.hpp>
+#include <metrics/metrics.hpp>
 
 /**
  * @file context.hpp defines WorkloadContext, ActorContext, and PhaseContext.
